+++ app/api/v1/feed.py
@@ -35,6 +35,7 @@ def list_feed_events():
         giveaway_distance_explicit=distance_explicit,
         included_event_types=query_data["types"],
         include_own_activity=query_data["show_own_activity"],
+        include_claimed_giveaways=query_data["show_claimed_giveaways"],
         max_events=None,
     )
     pagination = ListPagination(